Standardize ProfitYear type; add check processing policy
- Introduce IProfitYearRequest interface for consistent ProfitYear usage (short instead of int) in check run DTOs and services
- Update CheckRunStartRequest and CheckRunWorkflowResponse to implement IProfitYearRequest and use short ProfitYear
- Refactor service and orchestrator interfaces/implementations to accept short profitYear
- Add CanProcessChecks policy and map authorized roles in PolicyRoleMap
- Apply FtpOperationLogConfiguration in ContextExtensions for FTP log entity mapping
- Minor code cleanups and improved formatting for consistency

+++ src/services/src/Demoulas.ProfitSharing.Services/CheckRun/CheckRunOrchestrator.cs
@@ -32,8 +32,7 @@ public sealed class CheckRunOrchestrator : ICheckRunOrchestrator
     /// TODO: Integrate with IMicrFileGenerator, IDjdeFileGenerator, IPositivePayFileGenerator, and IFileTransferService
     /// when Todos #1-4 are fully implemented.
     /// </summary>
-    public Task<Result<bool>> ExecuteCheckRunAsync(
-        int profitYear,
+    public Task<Result<bool>> ExecuteCheckRunAsync(short profitYear,
         string checkNumber,
         Guid userId,
         Guid runId,
